frugen: Add support for internal use area creation
The internal use area can now be created from a JSON file.
No support for command line overrides.

To create an internal use area, specify an 'internal' attribute
followed by a hex string value. See the updated example.json file.
The field size is calculated automatically, so you need to
specify full default data contents for the area. The size of
the area will be automatically rounded up to a multiple of 8 bytes.

Signed-off-by: Alexander Amelkin <a.amelkin@yadro.com>